Move create/increase/decrease/redeem hooks to files
