+++ src/components/PositionModal/BorrowForms/CreateForm.tsx
@@ -1,7 +1,7 @@
 import { decToScale, wadToDec, ZERO } from '@fiatdao/sdk';
 import { Button, Card, Grid, Input, Loading, Modal, Row, Spacer, Switch, Text, Tooltip } from '@nextui-org/react';
-import { ethers } from 'ethers';
-import React, { useMemo } from 'react';
+import { BigNumber, ethers } from 'ethers';
+import React, { useCallback, useMemo } from 'react';
 import shallow from 'zustand/shallow';
 import 'katex/dist/katex.min.css';
 import { InlineMath } from 'react-katex';
@@ -11,8 +11,12 @@ import { commifyToDecimalPlaces, floor2, floor4, minCollRatioWithBuffer } from '
 import { Alert } from '../../Alert';
 import { NumericInput } from '../../NumericInput/NumericInput';
 import { Slider } from '../../Slider/Slider';
-import { useCreatePosition } from '../../../hooks/useBorrowPositions';
 import { useSetUnderlierAllowanceForProxy, useUnsetUnderlierAllowanceForProxy } from '../../../hooks/useSetAllowance';
+import { buildBuyCollateralAndModifyDebtArgs, sendTransaction } from '../../../actions';
+import { useAddRecentTransaction } from '@rainbow-me/rainbowkit';
+import { chain as chains, useAccount, useNetwork } from 'wagmi';
+import useSoftReset from '../../../hooks/useSoftReset';
+import { useUserData } from '../../../state/queries/useUserData';
 
 const CreateForm = ({
   onClose,
@@ -34,14 +38,33 @@ const CreateForm = ({
   );
   const modifyPositionData = useStore((state) => state.modifyPositionData);
   const fiat = useStore(state => state.fiat);
+  const user = useStore((state) => state.user);
   const hasProxy = useStore(state => state.hasProxy);
   const disableActions = useStore((state) => state.disableActions);
   const transactionData = useStore((state => state.transactionData));
   const { action: currentTxAction } = transactionData;
 
-  const createPosition = useCreatePosition();
   const setUnderlierAllowanceForProxy = useSetUnderlierAllowanceForProxy();
   const unsetUnderlierAllowanceForProxy = useUnsetUnderlierAllowanceForProxy();
+
+  const { chain } = useNetwork();
+  const { address } = useAccount();
+  const addRecentTransaction = useAddRecentTransaction();
+  const softReset = useSoftReset();
+
+  const { data: userData } = useUserData(fiat, chain?.id ?? chains.mainnet.id, address ?? '');
+  const { proxies } = userData as any;
+
+  const createPosition = useCallback(async (deltaCollateral: BigNumber, deltaDebt: BigNumber, underlier: BigNumber) => {
+    const args = buildBuyCollateralAndModifyDebtArgs(
+      fiat, user, proxies, modifyPositionData.collateralType, deltaCollateral, deltaDebt, underlier
+    );
+    const response = await sendTransaction(
+      fiat, true, proxies[0], 'createPosition', args.contract, args.methodName, ...args.methodArgs
+    );
+    addRecentTransaction({ hash: response.transactionHash, description: 'Create position' });
+    softReset();
+  }, [addRecentTransaction, fiat, modifyPositionData.collateralType, proxies, softReset, user]);
 
   const {
     collateralType: {
