+++ src/components/PositionModal/LeverForms/LeverCreateForm.tsx
@@ -1,7 +1,7 @@
 import { decToScale, interestPerSecondToAnnualYield, scaleToDec, wadToDec, ZERO } from '@fiatdao/sdk';
 import { Button, Card, Grid, Input, Loading, Modal, Row, Spacer, Switch, Text, Tooltip } from '@nextui-org/react';
-import { ethers } from 'ethers';
-import React, { useMemo } from 'react';
+import { BigNumber, ethers } from 'ethers';
+import React, { useCallback, useMemo } from 'react';
 import 'katex/dist/katex.min.css';
 import { InlineMath } from 'react-katex';
 import shallow from 'zustand/shallow';
@@ -11,8 +11,12 @@ import { commifyToDecimalPlaces, floor2, floor4 } from '../../../utils';
 import { Alert } from '../../Alert';
 import { NumericInput } from '../../NumericInput/NumericInput';
 import { Slider } from '../../Slider/Slider';
-import { useCreateLeveredPosition } from '../../../hooks/useLeveredPositions';
 import { useSetUnderlierAllowanceForProxy, useUnsetUnderlierAllowanceForProxy } from '../../../hooks/useSetAllowance';
+import { buildBuyCollateralAndIncreaseLeverArgs, sendTransaction } from '../../../actions';
+import { useAddRecentTransaction } from '@rainbow-me/rainbowkit';
+import { chain as chains, useAccount, useNetwork } from 'wagmi';
+import useSoftReset from '../../../hooks/useSoftReset';
+import { useUserData } from '../../../state/queries/useUserData';
 
 const LeverCreateForm = ({
   onClose,
@@ -33,14 +37,35 @@ const LeverCreateForm = ({
   );
 
   const fiat = useStore(state => state.fiat);
+  const user = useStore((state) => state.user);
   const hasProxy = useStore(state => state.hasProxy);
   const disableActions = useStore((state) => state.disableActions);
   const transactionData = useStore((state => state.transactionData));
   const modifyPositionData = useStore((state) => state.modifyPositionData);
 
-  const createLeveredPosition = useCreateLeveredPosition();
   const setUnderlierAllowanceForProxy = useSetUnderlierAllowanceForProxy();
   const unsetUnderlierAllowanceForProxy = useUnsetUnderlierAllowanceForProxy();
+
+  const addRecentTransaction = useAddRecentTransaction();
+  const { chain } = useNetwork();
+  const { address } = useAccount();
+  const softReset = useSoftReset();
+
+  const { data: userData } = useUserData(fiat, chain?.id ?? chains.mainnet.id, address ?? '');
+  const { proxies } = userData as any;
+  
+  const createLeveredPosition = useCallback(async (
+    upFrontUnderlier: BigNumber, addDebt: BigNumber, minUnderlierToBuy: BigNumber, minTokenToBuy: BigNumber
+  ) => {
+    const args = await buildBuyCollateralAndIncreaseLeverArgs(
+      fiat, user, proxies, modifyPositionData.collateralType, upFrontUnderlier, addDebt, minUnderlierToBuy, minTokenToBuy
+    );
+    const response = await sendTransaction(
+      fiat, true, proxies[0], 'createLeveredPosition', args.contract, args.methodName, ...args.methodArgs
+    );
+    addRecentTransaction({ hash: response.transactionHash, description: 'Create levered position' });
+    softReset();
+  }, [addRecentTransaction, fiat, modifyPositionData.collateralType, proxies, softReset, user]);
 
   const [submitError, setSubmitError] = React.useState('');
